requirements.txt
albumentations==1.3.0
imagesize==1.4.1
matplotlib==3.5.2
Pillow==9.0.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
kaggle==1.5.12
tqdm==4.64.1
numpy==1.21.2
pandas==1.4.3
scikit-learn==1.1.1
torch==1.11.0
torchvision==0.12.0
omegaconf==2.2.3
timm==0.6.6
faiss-gpu==1.7.2
open-clip-torch==2.0.2
tensorflow==2.10.0
pytorch-lightning==1.5.10
pytorch-metric-learning==1.6.2
-e git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup.git@2958117072f214dce743f918d6a2253f682ab65d#egg=cosine-annealing-warmup